Adding new serialization from JsonMessageFormat to GenericJackson...
